requirements.txt
asgiref==3.6.0
bleach==5.0.1
Django==4.1.4
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-api-logger==1.1.11
psycopg2==2.9.5
PyJWT==2.6.0
pytz==2022.7
six==1.16.0
sqlparse==0.4.3
webencodings==0.5.1
